feat: Finished interface exercise

+++ src/shapes/Measurable.java
@@ -1,6 +1,6 @@
 package shapes;
 
 interface Measurable {
-    double getPerimeter();
-    double getArea();
+    int getPerimeter();
+    int getArea();
 }
